+++ drools-compiler/src/main/java/org/drools/compiler/builder/impl/CompositeKnowledgeBuilderImpl.java
@@ -124,7 +124,6 @@ public class CompositeKnowledgeBuilderImpl implements CompositeKnowledgeBuilder 
         buildResourceType(ResourceBuilder.PKG_RESOURCE_BUILDER, ResourceType.PKG);
         buildResourceType(ResourceBuilder.CHANGE_SET_RESOURCE_BUILDER, ResourceType.CHANGE_SET);
         buildResourceType(ResourceBuilder.XSD_RESOURCE_BUILDER, ResourceType.XSD);
-        buildResourceType(ResourceBuilder.PMML_RESOURCE_BUILDER, ResourceType.PMML);
         buildResourceType(ResourceBuilder.SCD_RESOURCE_BUILDER, ResourceType.SCARD);
         buildResourceType(ResourceBuilder.GSCD_RESOURCE_BUILDER, ResourceType.SCGD);
     }
@@ -284,7 +283,6 @@ public class CompositeKnowledgeBuilderImpl implements CompositeKnowledgeBuilder 
 
         ResourceBuilder DSL_RESOURCE_BUILDER = ( kBuilder, resourceDescr ) -> kBuilder.addDsl( resourceDescr.resource );
 
-        ResourceBuilder PMML_RESOURCE_BUILDER = ( kBuilder, resourceDescr ) -> kBuilder.addPackageFromPMML(resourceDescr.resource, ResourceType.PMML, resourceDescr.configuration);
 
         ResourceBuilder XSD_RESOURCE_BUILDER = ( kBuilder, resourceDescr ) -> {
             if (resourceDescr.configuration instanceof JaxbConfigurationImpl) {
