+++ drools-compiler/src/test/java/org/drools/compiler/integrationtests/KnowledgeBuilderTest.java
@@ -15,16 +15,18 @@
 
 package org.drools.compiler.integrationtests;
 
-import java.io.InputStream;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
+
 import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
-import org.drools.compiler.compiler.PMMLCompiler;
-import org.drools.compiler.compiler.PMMLResource;
 import org.drools.core.definitions.impl.KnowledgePackageImpl;
 import org.drools.core.definitions.rule.impl.RuleImpl;
 import org.drools.core.impl.InternalKnowledgeBase;
@@ -36,11 +38,9 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
-import org.kie.api.builder.model.KieBaseModel;
 import org.kie.api.definition.KiePackage;
 import org.kie.api.definition.rule.Rule;
 import org.kie.api.definition.type.FactType;
-import org.kie.api.internal.utils.ServiceDiscoveryImpl;
 import org.kie.api.internal.utils.ServiceRegistry;
 import org.kie.api.internal.utils.ServiceRegistryImpl;
 import org.kie.api.io.Resource;
@@ -48,10 +48,7 @@ import org.kie.api.io.ResourceType;
 import org.kie.api.runtime.KieSession;
 import org.kie.internal.builder.KnowledgeBuilder;
 import org.kie.internal.builder.KnowledgeBuilderFactory;
-import org.kie.internal.builder.KnowledgeBuilderResult;
 import org.kie.internal.io.ResourceFactory;
-
-import static org.junit.Assert.*;
 
 public class KnowledgeBuilderTest {
 
@@ -459,7 +456,7 @@ public class KnowledgeBuilderTest {
     public void testResourceMapping() throws Exception {
         ServiceRegistryImpl serviceRegistry = (ServiceRegistryImpl) ServiceRegistry.getInstance();
         serviceRegistry.reset();
-
+/*
         ServiceDiscoveryImpl.getInstance().addService(PMMLCompiler.class.getCanonicalName(),
                                                       new PMMLCompiler() {
                                                           public String compile(InputStream stream, ClassLoader cl) {
@@ -506,7 +503,7 @@ public class KnowledgeBuilderTest {
 															return Collections.emptyMap();
 														}
                                                       } );
-
+*/
         serviceRegistry.reload();
 
         String rule = "package org.drools.compiler.test\n" +
