+++ kie-pmml/src/main/resources/META-INF/kie.conf
@@ -1,1 +1,1 @@
-org.drools.compiler.compiler.PMMLCompiler = org.kie.pmml.pmml_4_2.PMML4Compiler
+org.kie.api.internal.assembler.KieAssemblers = +org.kie.pmml.assembler.PMMLAssemblerService
